+++ lib/stripe_event/exceptions.rb
@@ -1,0 +1,5 @@
+module StripeEvent
+  Error = Class.new(StandardError)
+
+  UnauthorizedError = Class.new(Error)
+end
