+++ lib/stripe_event/notification_adapter.rb
@@ -1,0 +1,18 @@
+module StripeEvent
+  class NotificationAdapter
+    attr_accessor :subscriber
+
+    def self.call(callable)
+      new(callable)
+    end
+
+    def initialize(subscriber)
+      @subscriber = subscriber
+    end
+
+    def call(*args)
+      payload = args.last
+      subscriber.call(payload)
+    end
+  end
+end
